Move node_agent into cmd/ directory
Also inline `getTLSCredentials` into `onprem.go` since it's only used
there.

/cc @wattli @myidpt
