+++ cmd/node_agent/main.go
@@ -18,7 +18,7 @@ import (
 	"flag"
 
 	"github.com/golang/glog"
-	"istio.io/auth/node_agent/na"
+	"istio.io/auth/cmd/node_agent/na"
 )
 
 var (
