+++ cmd/node_agent/na/BUILD
@@ -14,10 +14,10 @@ go_library(
         "//pkg/pki:go_default_library",
         "//pkg/pki/ca:go_default_library",
         "//proto:go_default_library",
-        "//utils:go_default_library",
         "@com_github_golang_glog//:go_default_library",
         "@com_google_cloud_go//compute/metadata:go_default_library",
         "@org_golang_google_grpc//:go_default_library",
+        "@org_golang_google_grpc//credentials:go_default_library",
         "@org_golang_x_net//context:go_default_library",
     ],
 )
