+++ cmd/node_agent/na/gcp.go
