+++ cmd/node_agent/na/gcp_test.go
