--- utils/BUILD
@@ -1,11 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
-
-go_library(
-    name = "go_default_library",
-    srcs = ["tls.go"],
-    visibility = ["//visibility:public"],
-    deps = [
-        "@com_github_golang_glog//:go_default_library",
-        "@org_golang_google_grpc//credentials:go_default_library",
-    ],
-)
